// src/gen/proto/register/v1/register.ts
export interface RegisterDataNodeRequest {
  node_id: string;
  hostname: string;
  port: number;
  scheme: string;
  health_snapshot: HealthSnapshot | undefined;
}

export interface RegisterDataNodeResponse {
}

function createBaseRegisterDataNodeRequest(): RegisterDataNodeRequest {
  return { node_id: "", hostname: "", port: 0, scheme: "", health_snapshot: undefined };
}

export const RegisterDataNodeRequest: MessageFns<RegisterDataNodeRequest> = {
  encode(message: RegisterDataNodeRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.node_id !== "") {
      writer.uint32(10).string(message.node_id);
    }
    if (message.hostname !== "") {
      writer.uint32(18).string(message.hostname);
    }
    if (message.port !== 0) {
      writer.uint32(24).uint32(message.port);
    }
    if (message.scheme !== "") {
      writer.uint32(34).string(message.scheme);
    }
    if (message.health_snapshot !== undefined) {
      HealthSnapshot.encode(message.health_snapshot, writer.uint32(42).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): RegisterDataNodeRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRegisterDataNodeRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.node_id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.hostname = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.port = reader.uint32();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.scheme = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.health_snapshot = HealthSnapshot.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): RegisterDataNodeRequest {
    return {
      node_id: isSet(object.nodeId)
        ? globalThis.String(object.nodeId)
        : isSet(object.node_id)
        ? globalThis.String(object.node_id)
        : "",
      hostname: isSet(object.hostname) ? globalThis.String(object.hostname) : "",
      port: isSet(object.port) ? globalThis.Number(object.port) : 0,
      scheme: isSet(object.scheme) ? globalThis.String(object.scheme) : "",
      health_snapshot: isSet(object.healthSnapshot)
        ? HealthSnapshot.fromJSON(object.healthSnapshot)
        : isSet(object.health_snapshot)
        ? HealthSnapshot.fromJSON(object.health_snapshot)
        : undefined,
    };
  },

  toJSON(message: RegisterDataNodeRequest): unknown {
    const obj: any = {};
    if (message.node_id !== "") {
      obj.nodeId = message.node_id;
    }
    if (message.hostname !== "") {
      obj.hostname = message.hostname;
    }
    if (message.port !== 0) {
      obj.port = Math.round(message.port);
    }
    if (message.scheme !== "") {
      obj.scheme = message.scheme;
    }
    if (message.health_snapshot !== undefined) {
      obj.healthSnapshot = HealthSnapshot.toJSON(message.health_snapshot);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<RegisterDataNodeRequest>, I>>(base?: I): RegisterDataNodeRequest {
    return RegisterDataNodeRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<RegisterDataNodeRequest>, I>>(object: I): RegisterDataNodeRequest {
    const message = createBaseRegisterDataNodeRequest();
    message.node_id = object.node_id ?? "";
    message.hostname = object.hostname ?? "";
    message.port = object.port ?? 0;
    message.scheme = object.scheme ?? "";
    message.health_snapshot = (object.health_snapshot !== undefined && object.health_snapshot !== null)
      ? HealthSnapshot.fromPartial(object.health_snapshot)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}