// src/gen/proto/heartbeat/v1/heartbeat.ts
export interface HeartbeatDataNodeRequest {
  node_id: string;
  health_snapshot: HealthSnapshot | undefined;
}

export interface HeartbeatDataNodeResponse {
}

function createBaseHeartbeatDataNodeRequest(): HeartbeatDataNodeRequest {
  return { node_id: "", health_snapshot: undefined };
}

export const HeartbeatDataNodeRequest: MessageFns<HeartbeatDataNodeRequest> = {
  encode(message: HeartbeatDataNodeRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.node_id !== "") {
      writer.uint32(10).string(message.node_id);
    }
    if (message.health_snapshot !== undefined) {
      HealthSnapshot.encode(message.health_snapshot, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): HeartbeatDataNodeRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseHeartbeatDataNodeRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.node_id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.health_snapshot = HealthSnapshot.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): HeartbeatDataNodeRequest {
    return {
      node_id: isSet(object.nodeId)
        ? globalThis.String(object.nodeId)
        : isSet(object.node_id)
        ? globalThis.String(object.node_id)
        : "",
      health_snapshot: isSet(object.healthSnapshot)
        ? HealthSnapshot.fromJSON(object.healthSnapshot)
        : isSet(object.health_snapshot)
        ? HealthSnapshot.fromJSON(object.health_snapshot)
        : undefined,
    };
  },

  toJSON(message: HeartbeatDataNodeRequest): unknown {
    const obj: any = {};
    if (message.node_id !== "") {
      obj.nodeId = message.node_id;
    }
    if (message.health_snapshot !== undefined) {
      obj.healthSnapshot = HealthSnapshot.toJSON(message.health_snapshot);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<HeartbeatDataNodeRequest>, I>>(base?: I): HeartbeatDataNodeRequest {
    return HeartbeatDataNodeRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<HeartbeatDataNodeRequest>, I>>(object: I): HeartbeatDataNodeRequest {
    const message = createBaseHeartbeatDataNodeRequest();
    message.node_id = object.node_id ?? "";
    message.health_snapshot = (object.health_snapshot !== undefined && object.health_snapshot !== null)
      ? HealthSnapshot.fromPartial(object.health_snapshot)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}